+++ src/main/java/ar/edu/um/ingenieria/controller/foro/TemaController.java
@@ -1,9 +1,8 @@
 package ar.edu.um.ingenieria.controller.foro;
 
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.util.Date;
+import java.util.Calendar;
 import java.util.List;
+import java.util.TimeZone;
 
 import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -49,7 +48,7 @@ public class TemaController {
 			Categoria categoria = categoriaServiceImpl.findById(id);
 			if (categoria.getTemas() == null) {
 				logger.info("No hay temas en la categoria de ID:" + id);
-				new ResponseEntity<List<Tema>>(HttpStatus.NO_CONTENT);
+				return new ResponseEntity<List<Tema>>(HttpStatus.NO_CONTENT);
 			}
 			logger.info("Datos tema:" + categoria.getTemas());
 			return new ResponseEntity<List<Tema>>(categoria.getTemas(), HttpStatus.OK);
@@ -67,40 +66,52 @@ public class TemaController {
 	}
 
 	@PostMapping
-	public ResponseEntity<Void> insert(String titulo, Boolean cerrado, String texto, Integer idCategoria, String fecha)
-			throws ParseException {
-		SimpleDateFormat simpleDateFormat = new SimpleDateFormat("dd-MM-yyyy HH:mm:ss");
+	public ResponseEntity<Void> insert(String titulo, String texto, Integer idCategoria) {
+		if (categoriaServiceImpl.findById(idCategoria) == null) {
+			logger.info("No existe la categoria de ID:" + idCategoria);
+			return new ResponseEntity<Void>(HttpStatus.BAD_REQUEST);
+		}
+		Calendar calendar = Calendar.getInstance(TimeZone.getTimeZone("GMT-0300"));
+		calendar.add(Calendar.HOUR, -3);
 		Tema tema = new Tema();
 		tema.setTitulo(titulo);
 		tema.setUsuario(usuarioServiceImpl.findById(usuarioSecurity.GetIdUser()));
-		tema.setCerrado(cerrado);
+		tema.setCerrado(false);
 		tema.setTexto(texto);
 		tema.setCategoria(categoriaServiceImpl.findById(idCategoria));
-		Date date = simpleDateFormat.parse(fecha);
-		tema.setFecha(date);
+		tema.setFecha(calendar.getTime());
 		temaServiceImpl.create(tema);
 		logger.info("Tema creado con exito:" + tema);
 		return new ResponseEntity<Void>(HttpStatus.CREATED);
 	}
 
 	@PostMapping("/edit")
-	public ResponseEntity<Void> edit(String titulo, Boolean cerrado, String texto,
-			Integer idCategoria, String fecha, Integer id) throws ParseException {
+	public ResponseEntity<Void> edit(String titulo, Boolean cerrado, String texto, Integer idCategoria, Integer id) {
 		if (temaServiceImpl.findById(id) == null) {
 			logger.info("No existe el tema de ID:" + id);
 			return new ResponseEntity<Void>(HttpStatus.NO_CONTENT);
 		}
-		SimpleDateFormat simpleDateFormat = new SimpleDateFormat("dd/MM/yyyy HH:mm:ss");
-		Tema tema = temaServiceImpl.findById(id);
-		tema.setTitulo(titulo);
-		tema.setUsuario(usuarioServiceImpl.findById(usuarioSecurity.GetIdUser()));
-		tema.setCerrado(cerrado);
-		tema.setTexto(texto);
-		tema.setCategoria(categoriaServiceImpl.findById(idCategoria));
-		tema.setFecha(simpleDateFormat.parse(fecha));
-		temaServiceImpl.create(tema);
-		logger.info("Tema actualizado con exito:" + tema);
-		return new ResponseEntity<Void>(HttpStatus.CREATED);
+		if (categoriaServiceImpl.findById(idCategoria) == null) {
+			logger.info("No existe la categoria de ID:" + idCategoria);
+			return new ResponseEntity<Void>(HttpStatus.BAD_REQUEST);
+		}
+		if (temaServiceImpl.findById(id).getUsuario() == usuarioServiceImpl.findById(usuarioSecurity.GetIdUser())) {
+			Calendar calendar = Calendar.getInstance(TimeZone.getTimeZone("GMT-0300"));
+			calendar.add(Calendar.HOUR, -3);
+			Tema tema = temaServiceImpl.findById(id);
+			tema.setTitulo(titulo);
+			tema.setUsuario(usuarioServiceImpl.findById(usuarioSecurity.GetIdUser()));
+			tema.setCerrado(cerrado);
+			tema.setTexto(texto);
+			tema.setCategoria(categoriaServiceImpl.findById(idCategoria));
+			tema.setFecha(calendar.getTime());
+			temaServiceImpl.create(tema);
+			logger.info("Tema actualizado con exito:" + tema);
+			return new ResponseEntity<Void>(HttpStatus.CREATED);
+		}
+		logger.info("Tema no actualizado. El usuario no es propietario del tema.");
+		return new ResponseEntity<Void>(HttpStatus.BAD_REQUEST);
+
 	}
 
 	@DeleteMapping("/{id}")
@@ -109,8 +120,12 @@ public class TemaController {
 			logger.info("No existe el tema de ID:" + id);
 			return new ResponseEntity<Void>(HttpStatus.NO_CONTENT);
 		}
-		logger.info("Tema borrado con exito:" + temaServiceImpl.findById(id));
-		temaServiceImpl.remove(temaServiceImpl.findById(id));
-		return new ResponseEntity<Void>(HttpStatus.OK);
+		if (temaServiceImpl.findById(id).getUsuario() == usuarioServiceImpl.findById(usuarioSecurity.GetIdUser())) {
+			logger.info("Tema borrado con exito:" + temaServiceImpl.findById(id));
+			temaServiceImpl.remove(temaServiceImpl.findById(id));
+			return new ResponseEntity<Void>(HttpStatus.OK);
+		}
+		logger.info("Tema no borrado. El usuario no es propietario del tema.");
+		return new ResponseEntity<Void>(HttpStatus.BAD_REQUEST);
 	}
 }
